+++ src/account/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
 
-class AccountConfig(AppConfig):
+class AccountsConfig(AppConfig):
     name = 'account'
