+++ src/resource/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class StudyMaterialConfig(AppConfig):
-    name = 'study_material'
+    name = 'resource'
